Extract ReaderAddSubscriptionButton
