+++ WordPress/Classes/ViewRelated/Reader/Subscriptions/ReaderSubscriptionsView.swift
@@ -9,7 +9,7 @@ struct ReaderSubscriptionsView: View {
     private var subscriptions: FetchedResults<ReaderSiteTopic>
 
     @State private var searchText = ""
-    @State private var isAddingSubscription = false
+    @State private var isShowingMainAddSubscriptonPopover = false
 
     @State private var searchResults: [ReaderSiteTopic] = []
 
@@ -42,6 +42,12 @@ struct ReaderSubscriptionsView: View {
         .refreshable {
             await viewModel.refresh()
         }
+        .toolbar {
+            ReaderAddSubscriptionButton(style: .navigation)
+            if !subscriptions.isEmpty {
+                EditButton()
+            }
+        }
         .navigationTitle(Strings.title)
     }
 
@@ -64,18 +70,14 @@ struct ReaderSubscriptionsView: View {
         } description: {
             Text(Strings.emptyStateDetails)
         } actions: {
-            buttonAddSubscription
-                .fixedSize(horizontal: true, vertical: /*@START_MENU_TOKEN@*/true/*@END_MENU_TOKEN@*/)
-                .padding(EdgeInsets(top: 12, leading: 8, bottom: 8, trailing: 10))
-                .background(Color(.secondarySystemBackground).opacity(0.8))
-                .clipShape(RoundedRectangle(cornerRadius: 12))
+            ReaderAddSubscriptionButton(style: .compact)
         }
     }
 
     private var main: some View {
         List {
             if !isShowingSearchResuts {
-                buttonAddSubscription
+                ReaderAddSubscriptionButton(style: .expanded)
                     .listRowSeparator(.hidden)
                 ForEach(subscriptions, id: \.objectID, content: makeSubscriptionCell)
                     .onDelete(perform: delete)
@@ -94,9 +96,6 @@ struct ReaderSubscriptionsView: View {
         .onChange(of: searchText) {
             reloadSearchResults(searchText: $0)
         }
-        .toolbar {
-            EditButton()
-        }
     }
 
     private func makeSubscriptionCell(for site: ReaderSiteTopic) -> some View {
@@ -104,33 +103,6 @@ struct ReaderSubscriptionsView: View {
             onSelection(site)
         } label: {
             ReaderSubscriptionCell(site: site, onDelete: delete)
-        }
-    }
-
-    private var buttonAddSubscription: some View {
-        Button {
-            isAddingSubscription = true
-        } label: {
-            HStack(spacing: 16) {
-                Image(systemName: "plus.circle.fill")
-                    .foregroundStyle(AppColor.brand, Color(.secondarySystemFill))
-                    .font(.largeTitle.weight(.light))
-                    .frame(width: 40)
-                    .padding(.leading, 4)
-                VStack(alignment: .leading) {
-                    Text(Strings.addSubscription)
-                        .font(.callout.weight(.medium))
-                    Text(Strings.addSubscriptionSubtitle)
-                        .font(.footnote)
-                        .foregroundStyle(.secondary)
-                }
-                .lineLimit(1)
-            }
-            .padding(.bottom, 4)
-        }
-        .buttonStyle(.plain)
-        .popover(isPresented: $isAddingSubscription) {
-            ReaderAddSubscriptionView()
         }
     }
 
@@ -158,52 +130,6 @@ struct ReaderSubscriptionsView: View {
     }
 }
 
-private struct ReaderAddSubscriptionView: View {
-    @State private var url = ""
-    @FocusState private var isFocused: Bool
-    @Environment(\.dismiss) private var dismiss
-
-    var body: some View {
-        VStack(alignment: .leading, spacing: 12) {
-            controls
-            Text(Strings.addSubscriptionSubtitle)
-                .font(.footnote)
-                .foregroundStyle(.secondary)
-            TextField("", text: $url, prompt: Text(verbatim: "example.com"))
-                .keyboardType(.URL)
-                .textContentType(.URL)
-                .textInputAutocapitalization(.never)
-                .autocorrectionDisabled()
-                .focused($isFocused)
-                .labelsHidden()
-                .padding(.top, 12)
-        }
-        .padding()
-        .onAppear {
-            isFocused = true
-        }
-        .frame(width: 420)
-        .interactiveDismissDisabled(!url.isEmpty)
-    }
-
-    private var controls: some View {
-        HStack {
-            Button(SharedStrings.Button.cancel) {
-                dismiss()
-            }
-            Spacer()
-            Text(Strings.addSubscription)
-                .font(.headline)
-            Spacer()
-            Button(SharedStrings.Button.add) {
-                // TODO: implement
-            }
-            .disabled(url.isEmpty)
-            .font(.headline)
-        }
-    }
-}
-
 enum ReaderSubscriptionHelper {
     static func unfollow(_ site: ReaderSiteTopic, in context: CoreDataStackSwift = ContextManager.shared) {
         NotificationCenter.default.post(name: .ReaderTopicUnfollowed, object: nil, userInfo: [ReaderNotificationKeys.topic: site])
@@ -220,7 +146,5 @@ enum ReaderSubscriptionHelper {
 
 private enum Strings {
     static let title = NSLocalizedString("reader.subscriptions.title", value: "Subscriptions", comment: "Navigation bar title")
-    static let addSubscription = NSLocalizedString("reader.subscriptions.addSubscriptionButtonTitle", value: "Add Subscription", comment: "Button title")
-    static let addSubscriptionSubtitle = NSLocalizedString("reader.subscriptions.addSubscriptionButtonSubtitle", value: "Subscribe to sites, newsletters, or RSS feeds", comment: "Button subtitle")
     static let emptyStateDetails = NSLocalizedString("reader.subscriptions.emptyStateDetails", value: "The sites you discover and subscribe to will appear here", comment: "Empty state details")
 }
